[vfs] E: Extract Descriptor Handles
Move backend-neutral descriptor handles into dedicated one-type modules
and re-export them through fd to preserve the public API. Redirect the
FAT32 adapter to the descriptor layer so it no longer depends on
FD-table implementation details.
