+++ src/libs/vfs/src/descriptor/direct_read_handle.rs
@@ -1,0 +1,85 @@
+// Copyright(c) The Maintainers of Nanvix.
+// Licensed under the MIT License.
+
+//! Zero-copy direct-read descriptor handle.
+
+//==================================================================================================
+// Imports
+//==================================================================================================
+
+use ::fat32::Fat32Error;
+use ::sysapi::{
+    ffi::c_int,
+    sys_types::off_t,
+    unistd::file_seek,
+};
+
+//==================================================================================================
+// Structures
+//==================================================================================================
+
+/// Zero-copy direct memory access handle for file reads.
+///
+/// When a file's data is stored contiguously in an in-memory filesystem
+/// image, reads can be served directly from the image buffer via memcpy,
+/// bypassing all cluster chain traversal.
+pub struct DirectReadHandle {
+    /// Pointer to the file's data within the filesystem image.
+    data: *const u8,
+    /// File size in bytes.
+    size: usize,
+    /// Current read position.
+    position: usize,
+}
+
+//==================================================================================================
+// Implementations
+//==================================================================================================
+
+impl DirectReadHandle {
+    /// Creates a new direct read handle.
+    pub fn new(data: *const u8, size: usize) -> Self {
+        Self {
+            data,
+            size,
+            position: 0,
+        }
+    }
+
+    /// Reads data from the direct memory region.
+    pub fn read(&mut self, buf: &mut [u8]) -> usize {
+        let remaining: usize = self.size.saturating_sub(self.position);
+        let to_read: usize = buf.len().min(remaining);
+        if to_read == 0 {
+            return 0;
+        }
+        // SAFETY: data pointer is valid for the lifetime of the filesystem
+        // image, and position + to_read <= size (guaranteed by min above).
+        unsafe {
+            let src: *const u8 = self.data.add(self.position);
+            core::ptr::copy_nonoverlapping(src, buf.as_mut_ptr(), to_read);
+        }
+        self.position += to_read;
+        to_read
+    }
+
+    /// Seeks to a position in the direct memory region.
+    pub fn seek(&mut self, offset: off_t, whence: c_int) -> Result<off_t, Fat32Error> {
+        let new_pos: i64 = match whence {
+            file_seek::SEEK_SET => offset,
+            file_seek::SEEK_CUR => self.position as i64 + offset,
+            file_seek::SEEK_END => self.size as i64 + offset,
+            _ => return Err(Fat32Error::InvalidArgument),
+        };
+        if new_pos < 0 || new_pos > self.size as i64 {
+            return Err(Fat32Error::InvalidSeek);
+        }
+        self.position = new_pos as usize;
+        Ok(new_pos as off_t)
+    }
+
+    /// Returns the file size.
+    pub fn size(&self) -> usize {
+        self.size
+    }
+}
