+++ src/libs/vfs/src/fd.rs
@@ -95,415 +95,19 @@ const STAT_SECTOR_SIZE: u64 = 512;
 const DEFAULT_CWD: &str = "/";
 
 //==================================================================================================
-// Metadata
+// Public Re-Exports
 //==================================================================================================
 
-/// File metadata returned by stat operations.
-///
-/// This is the VFS-level metadata type, independent of any concrete
-/// filesystem. Backend modules translate their native metadata into this
-/// type.
-pub struct VfsStat {
-    /// File size in bytes (0 for directories).
-    size: u64,
-    /// Whether this entry is a directory.
-    is_dir: bool,
-}
-
-impl VfsStat {
-    /// Creates a new `VfsStat`.
-    pub fn new(size: u64, is_dir: bool) -> Self {
-        Self { size, is_dir }
-    }
-
-    /// Returns the file size in bytes.
-    pub fn size(&self) -> u64 {
-        self.size
-    }
-
-    /// Returns whether this entry is a directory.
-    pub fn is_dir(&self) -> bool {
-        self.is_dir
-    }
-}
-
-//==================================================================================================
-// Direct Read Handle
-//==================================================================================================
-
-/// Zero-copy direct memory access handle for file reads.
-///
-/// When a file's data is stored contiguously in an in-memory filesystem
-/// image, reads can be served directly from the image buffer via memcpy,
-/// bypassing all cluster chain traversal.
-pub struct DirectReadHandle {
-    /// Pointer to the file's data within the filesystem image.
-    data: *const u8,
-    /// File size in bytes.
-    size: usize,
-    /// Current read position.
-    position: usize,
-}
-
-impl DirectReadHandle {
-    /// Creates a new direct read handle.
-    pub fn new(data: *const u8, size: usize) -> Self {
-        Self {
-            data,
-            size,
-            position: 0,
-        }
-    }
-
-    /// Reads data from the direct memory region.
-    pub fn read(&mut self, buf: &mut [u8]) -> usize {
-        let remaining: usize = self.size.saturating_sub(self.position);
-        let to_read: usize = buf.len().min(remaining);
-        if to_read == 0 {
-            return 0;
-        }
-        // SAFETY: data pointer is valid for the lifetime of the filesystem
-        // image, and position + to_read <= size (guaranteed by min above).
-        unsafe {
-            let src: *const u8 = self.data.add(self.position);
-            core::ptr::copy_nonoverlapping(src, buf.as_mut_ptr(), to_read);
-        }
-        self.position += to_read;
-        to_read
-    }
-
-    /// Seeks to a position in the direct memory region.
-    pub fn seek(&mut self, offset: off_t, whence: c_int) -> Result<off_t, Fat32Error> {
-        let new_pos: i64 = match whence {
-            file_seek::SEEK_SET => offset,
-            file_seek::SEEK_CUR => self.position as i64 + offset,
-            file_seek::SEEK_END => self.size as i64 + offset,
-            _ => return Err(Fat32Error::InvalidArgument),
-        };
-        if new_pos < 0 || new_pos > self.size as i64 {
-            return Err(Fat32Error::InvalidSeek);
-        }
-        self.position = new_pos as usize;
-        Ok(new_pos as off_t)
-    }
-
-    /// Returns the file size.
-    pub fn size(&self) -> usize {
-        self.size
-    }
-}
-
-//==================================================================================================
-// VFS File Handle
-//==================================================================================================
-
-/// An open file handle managed by the VFS.
-///
-/// Each variant corresponds to a concrete filesystem backend or an
-/// optimization path. The VFS FD table stores these handles and
-/// dispatches operations to the appropriate variant.
-pub enum VfsFileHandle {
-    /// File opened through the FAT32 backend.
-    Fat32(crate::File),
-    /// Zero-copy direct memory read (contiguous file optimization).
-    DirectRead(DirectReadHandle),
-    /// Open directory handle for `readdir()`/`getdents()` operations.
-    Directory(DirectoryHandle),
-    /// Remote file opened through the host filesystem daemon (hostfsd).
-    /// Operations on this handle must be forwarded via IKC by the caller (vfsd).
-    HostFs(HostFsHandle),
-    /// One end of a POSIX unnamed pipe.
-    Pipe(crate::pipe::PipeEnd),
-    /// Routing token for a console stream (stdin/stdout/stderr).
-    ///
-    /// The handle lets vfsd own the descriptor slot, its per-descriptor flags, and the shared
-    /// console line discipline. The daemon still drives host I/O around this token.
-    Console(ConsoleHandle),
-    /// Routing token for a socket, holding the descriptor assigned by `networkd`.
-    ///
-    /// Socket I/O is not served by vfsd; like [`VfsFileHandle::HostFs`] this token only stores the
-    /// remote descriptor so vfsd can own the slot and its per-descriptor flags. vfsd closes the
-    /// remote descriptor on `networkd` when the last reference to the slot is dropped.
-    Socket(SocketHandle),
-}
-
-/// Handle for a file opened on the host filesystem via hostfsd.
-///
-/// This handle stores the remote file descriptor returned by hostfsd.
-/// The VFS cannot perform I/O on this handle directly — all operations
-/// must be forwarded via IKC by the owning daemon (vfsd).
-///
-/// The `is_dir` flag is set once at open time and never re-checked. If the
-/// host-side path changes type out-of-band (e.g., replaced by a directory),
-/// subsequent operations will use the stale classification.
-pub struct HostFsHandle {
-    /// Remote file descriptor on the host side.
-    remote_fd: i32,
-    /// Whether this is a directory.
-    is_dir: bool,
-    /// Absolute path used to open this handle (stored only for directories to support dirfd).
-    path: Option<String>,
-    /// Next directory entry index to return on the following `getdents` call.
-    ///
-    /// hostfsd serves directory listings via offset-based iteration (one entry per
-    /// `offset`), so vfsd tracks the per-FD cursor here. It advances as entries are
-    /// consumed and is meaningful only for directory handles.
-    readdir_offset: u32,
-}
-
-impl HostFsHandle {
-    /// Creates a new HostFs handle with the given remote file descriptor.
-    ///
-    /// The `path` argument is only meaningful for directory handles (used by dirfd resolution).
-    /// Pass `None` for regular file handles to avoid unnecessary allocations.
-    pub fn new(remote_fd: i32, is_dir: bool, path: Option<String>) -> Self {
-        Self {
-            remote_fd,
-            is_dir,
-            path: if is_dir { path } else { None },
-            readdir_offset: 0,
-        }
-    }
-
-    /// Returns the remote file descriptor.
-    pub fn remote_fd(&self) -> i32 {
-        self.remote_fd
-    }
-
-    /// Returns whether this is a directory handle.
-    pub fn is_dir(&self) -> bool {
-        self.is_dir
-    }
-
-    /// Returns the path used to open this handle (only available for directories).
-    pub fn path(&self) -> Option<&str> {
-        self.path.as_deref()
-    }
-
-    /// Returns the current directory iteration cursor.
-    pub fn readdir_offset(&self) -> u32 {
-        self.readdir_offset
-    }
-
-    /// Sets the directory iteration cursor.
-    pub fn set_readdir_offset(&mut self, offset: u32) {
-        self.readdir_offset = offset;
-    }
-}
-
-/// Identifies which standard console stream a [`ConsoleHandle`] represents.
-///
-/// A console-backed descriptor performs no I/O of its own; the stream identity is the only state it
-/// needs so that `fstat` can synthesize a stable device identity in a later plan.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum ConsoleStream {
-    /// Standard input (descriptor 0).
-    Stdin,
-    /// Standard output (descriptor 1).
-    Stdout,
-    /// Standard error (descriptor 2).
-    Stderr,
-}
-
-/// Routing token for a console-backed descriptor.
-///
-/// A console handle records which standard stream the descriptor represents and a reference to the
-/// shared line discipline (terminal attributes plus cooked input), but performs no I/O of its own.
-/// It exists so that vfsd can own the descriptor slot, its per-descriptor flags, and the terminal
-/// device state while the actual console I/O is driven from the daemon.
-pub struct ConsoleHandle {
-    /// Which standard stream this descriptor represents.
-    stream: ConsoleStream,
-    /// Shared line discipline (terminal attributes and cooked input), referenced by every console
-    /// descriptor of the device.
-    terminal: Arc<Mutex<LineDiscipline>>,
-}
-
-impl ConsoleHandle {
-    /// Creates a console handle for the given standard stream with its own fresh line discipline.
-    ///
-    /// Use [`ConsoleHandle::with_terminal`] when several handles must share one terminal device, as
-    /// the standard streams do.
-    pub fn new(stream: ConsoleStream) -> Self {
-        Self::with_terminal(stream, Arc::new(Mutex::new(LineDiscipline::default())))
-    }
-
-    /// Creates a console handle for the given standard stream that references `terminal`.
-    ///
-    /// The three standard streams are created this way from a single shared line discipline so that
-    /// they observe one consistent `termios`/`winsize` and one shared input buffer.
-    pub fn with_terminal(stream: ConsoleStream, terminal: Arc<Mutex<LineDiscipline>>) -> Self {
-        Self { stream, terminal }
-    }
-
-    /// Returns the standard stream this handle represents.
-    pub fn stream(&self) -> ConsoleStream {
-        self.stream
-    }
-
-    /// Returns a reference to the shared line discipline backing this console descriptor.
-    pub fn terminal(&self) -> &Arc<Mutex<LineDiscipline>> {
-        &self.terminal
-    }
-}
-
-/// Routing token for a socket-backed descriptor.
-///
-/// A socket handle stores the descriptor that `networkd` assigned to the socket (its remote fd),
-/// analogous to [`HostFsHandle::remote_fd`]. Socket I/O is not served by vfsd; this token only lets
-/// vfsd own the descriptor slot and its per-descriptor flags. vfsd closes the remote descriptor on
-/// `networkd` when the last reference to the slot is dropped.
-pub struct SocketHandle {
-    /// Descriptor assigned by `networkd` (the remote fd).
-    remote_fd: i32,
-}
-
-impl SocketHandle {
-    /// Creates a socket handle for the given `networkd` descriptor.
-    pub fn new(remote_fd: i32) -> Self {
-        Self { remote_fd }
-    }
-
-    /// Returns the `networkd` descriptor (remote fd) backing this socket.
-    pub fn remote_fd(&self) -> i32 {
-        self.remote_fd
-    }
-}
-
-/// Handle for an open directory.
-///
-/// Stores the resolved path and lazily-loaded directory entries.
-/// Entries are loaded on the first `getdents()` call and returned
-/// in subsequent calls via an internal cursor.
-pub struct DirectoryHandle {
-    /// Absolute path of the directory in the VFS.
-    path: String,
-    /// Cached directory entries (populated on first read).
-    entries: Option<Vec<crate::DirEntry>>,
-    /// Cursor into `entries` for sequential reads.
-    cursor: usize,
-}
-
-impl DirectoryHandle {
-    /// Creates a new directory handle for the given VFS path.
-    pub fn new(path: String) -> Self {
-        Self {
-            path,
-            entries: None,
-            cursor: 0,
-        }
-    }
-
-    /// Returns the next batch of directory entries.
-    ///
-    /// Lazily loads entries from the VFS on the first call and returns
-    /// up to `count` entries per invocation.
-    pub fn read_entries(&mut self, count: usize) -> Result<Vec<crate::DirEntry>, Fat32Error> {
-        if self.entries.is_none() {
-            self.entries = Some(crate::read_dir(&self.path)?);
-        }
-        let all: &[crate::DirEntry] = self.entries.as_ref().unwrap();
-        let remaining: &[crate::DirEntry] = if self.cursor < all.len() {
-            &all[self.cursor..]
-        } else {
-            &[]
-        };
-        let take: usize = core::cmp::min(count, remaining.len());
-        let batch: Vec<crate::DirEntry> = remaining[..take].to_vec();
-        self.cursor += take;
-        Ok(batch)
-    }
-}
-
-impl VfsFileHandle {
-    /// Reads data from the file.
-    pub fn read(&mut self, buf: &mut [u8]) -> Result<usize, Fat32Error> {
-        match self {
-            VfsFileHandle::Fat32(file) => file.read(buf),
-            VfsFileHandle::DirectRead(handle) => Ok(handle.read(buf)),
-            VfsFileHandle::Directory(_) => Err(Fat32Error::NotSupported),
-            VfsFileHandle::HostFs(_) => Err(Fat32Error::NotSupported),
-            // Pipes are served by vfsd through the dedicated non-blocking primitives.
-            VfsFileHandle::Pipe(_) => Err(Fat32Error::NotSupported),
-            // Console and socket tokens are inert routing markers; vfsd serves no I/O on them.
-            VfsFileHandle::Console(_) | VfsFileHandle::Socket(_) => Err(Fat32Error::NotSupported),
-        }
-    }
-
-    /// Writes data to the file.
-    pub fn write(&mut self, buf: &[u8]) -> Result<usize, Fat32Error> {
-        match self {
-            VfsFileHandle::Fat32(file) => file.write(buf),
-            VfsFileHandle::DirectRead(_) => Err(Fat32Error::ReadOnly),
-            VfsFileHandle::Directory(_) => Err(Fat32Error::NotSupported),
-            VfsFileHandle::HostFs(_) => Err(Fat32Error::NotSupported),
-            // Pipes are served by vfsd through the dedicated non-blocking primitives.
-            VfsFileHandle::Pipe(_) => Err(Fat32Error::NotSupported),
-            // Console and socket tokens are inert routing markers; vfsd serves no I/O on them.
-            VfsFileHandle::Console(_) | VfsFileHandle::Socket(_) => Err(Fat32Error::NotSupported),
-        }
-    }
-
-    /// Seeks to a position in the file.
-    pub fn seek(&mut self, offset: off_t, whence: c_int) -> Result<off_t, Fat32Error> {
-        match self {
-            VfsFileHandle::Fat32(file) => {
-                let pos: u64 = file.seek(whence, offset)?;
-                Ok(pos as off_t)
-            },
-            VfsFileHandle::DirectRead(handle) => handle.seek(offset, whence),
-            VfsFileHandle::Directory(_) => Err(Fat32Error::NotSupported),
-            VfsFileHandle::HostFs(_) => Err(Fat32Error::NotSupported),
-            // A pipe is not seekable (`ESPIPE`); the daemon rejects seeks before reaching here.
-            VfsFileHandle::Pipe(_) => Err(Fat32Error::NotSupported),
-            // Console and socket tokens are inert routing markers; vfsd serves no I/O on them.
-            VfsFileHandle::Console(_) | VfsFileHandle::Socket(_) => Err(Fat32Error::NotSupported),
-        }
-    }
-
-    /// Returns the file size in bytes.
-    pub fn size(&mut self) -> Result<u64, Fat32Error> {
-        match self {
-            VfsFileHandle::Fat32(file) => file.size(),
-            VfsFileHandle::DirectRead(handle) => Ok(handle.size() as u64),
-            VfsFileHandle::Directory(_) => Ok(0),
-            VfsFileHandle::HostFs(_) => Ok(0),
-            VfsFileHandle::Pipe(_) => Ok(0),
-            // Console and socket tokens have no size of their own.
-            VfsFileHandle::Console(_) | VfsFileHandle::Socket(_) => Ok(0),
-        }
-    }
-
-    /// Returns whether this handle is a directory.
-    pub fn is_dir(&self) -> bool {
-        match self {
-            VfsFileHandle::Directory(_) => true,
-            VfsFileHandle::HostFs(h) => h.is_dir(),
-            _ => false,
-        }
-    }
-
-    /// Returns whether this handle is backed by the host filesystem.
-    pub fn is_hostfs(&self) -> bool {
-        matches!(self, VfsFileHandle::HostFs(_))
-    }
-
-    /// Returns the remote FD if this is a HostFs handle.
-    pub fn hostfs_remote_fd(&self) -> Option<i32> {
-        match self {
-            VfsFileHandle::HostFs(h) => Some(h.remote_fd()),
-            _ => None,
-        }
-    }
-
-    /// Returns the pipe end if this handle is one end of a pipe.
-    pub fn pipe_end(&self) -> Option<&crate::pipe::PipeEnd> {
-        match self {
-            VfsFileHandle::Pipe(end) => Some(end),
-            _ => None,
-        }
-    }
-}
+pub use crate::descriptor::{
+    ConsoleHandle,
+    ConsoleStream,
+    DirectReadHandle,
+    DirectoryHandle,
+    HostFsHandle,
+    SocketHandle,
+    VfsFileHandle,
+    VfsStat,
+};
 
 //==================================================================================================
 // File Descriptor Table
@@ -1604,7 +1208,7 @@ pub fn vfs_resolve_path(dirfd: c_int, path: &str) -> Option<String> {
     let file: OpenFile = entry_arc(dirfd).ok()?;
     let guard = file.lock();
     let dir_path: &str = match &guard.handle {
-        VfsFileHandle::Directory(dh) => &dh.path,
+        VfsFileHandle::Directory(dh) => dh.path(),
         VfsFileHandle::HostFs(hh) if hh.is_dir() => hh.path()?,
         _ => return None, // fd is not a directory
     };
@@ -2101,7 +1705,7 @@ pub fn vfs_fchdir(fd: c_int) -> Result<(), Fat32Error> {
     let path: String = {
         let guard = file.lock();
         match &guard.handle {
-            VfsFileHandle::Directory(dir) => dir.path.clone(),
+            VfsFileHandle::Directory(dir) => String::from(dir.path()),
             _ => return Err(Fat32Error::NotADirectory),
         }
     };
